+++ libs/rxjs/array/src/lib/some.ts
@@ -6,6 +6,7 @@
 import { PredicateFn } from '../types/generic-methods';
 import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
+import { ArrayOrSet } from '../types/array-set';
 
 /**
  * Returns an Observable that emits a boolean when all values in the source Array or Set return truthy using Array.some
@@ -36,11 +37,11 @@ import { map } from 'rxjs/operators';
  *
  * @returns An Observable that emits a boolean when all values in source array return truthy
  */
-export function some<T extends unknown>(predicate?: PredicateFn<T>): OperatorFunction<T[] | Set<T>, boolean> {
+export function some<T extends unknown>(predicate?: PredicateFn<T>): OperatorFunction<ArrayOrSet<T>, boolean> {
   return (source) =>
     source.pipe(
-      map(([...value]) =>
-        value.some((v) => {
+      map((value) =>
+        [...value].some((v) => {
           if (predicate && typeof v === 'number') {
             return predicate(v);
           }