+++ libs/rxjs/array/src/lib/sort.ts
@@ -6,6 +6,7 @@ import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { defaultSortFn } from '../utils/sort';
 import { SortFn } from '../types/generic-methods';
+import { ArrayOrSet } from '../types/array-set';
 
 /**
  * Returns an Observable that emits an array of sorted values from the source Array or Set using the [[SortFn]]
@@ -42,6 +43,6 @@ import { SortFn } from '../types/generic-methods';
  *
  * @returns Observable array of values from source array sorted via [[SortFn]]
  */
-export function sort<T extends unknown>(sortFn?: SortFn<T>): OperatorFunction<T[] | Set<T>, T[]> {
+export function sort<T extends unknown>(sortFn?: SortFn<T>): OperatorFunction<ArrayOrSet<T>, T[]> {
   return (source) => source.pipe(map((value) => [...value].sort(sortFn || defaultSortFn)));
 }